// frontend-admin/src/api/generated/model/finanzOnlineMetricsResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface FinanzOnlineMetricsResponse {
  submitFailedPermanent?: number;
  submitFailedTotal?: number;
  submitFailedTransient?: number;
  submitFailedUnknown?: number;
  submitTotal?: number;
}